changes to homepageintro
