+++ src/Components/AboutUs.css
@@ -5,6 +5,7 @@
     background-color: #2525259e;
     border-radius: 12px;
     backdrop-filter: blur(100px);
+    height: 70vh;
     width: auto;
     padding: 40px;
 }
@@ -39,4 +40,12 @@
     color: white;
     font-size: 3rem;
     font-weight: 500;
+}
+
+.teacherImages {
+    height: 80px;
+    width: 80px;
+    object-fit: cover;
+    border-radius: 50%;
+    margin-top: -30px;
 }
